pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Denys Mitlenko <mitlenkodenis3@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
django-widget-tweaks = "^1.5.0"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
django-allauth = "^65.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.pyi?$'
force-exclude = '''
/(
  | migrations
)/
'''
